Folder: Generate an alias for new folders
Before commit d1f75b12e4309fa3308cfb717d682a7bbe1f8660, the wizard was
making sure folder had an alias but this is no longer the case.
So generate still an unique alias.

Alias is not used in the UI any longer, it's just use for internal purposes.

For issue #4737