Revert "Fix a deadlock when shutting down during discovery" (#5100)
Reverts commit db6ff5038cd8cbef5c3a0eb7ec4582d728106b30

Could be the cause of #5092 and the cost is higher than the benefit if this is the case.
A network request taking more than 30 seconds isn't something unlikely in this world
and shouldn't be a good reason to abort. We should try to untangle the threads
dependencies to properly fix this if possible instead.