Fix the Sphinx-doc build configuration
I'm not sure what the configuration/setup has been like to date. But as
it stood, it didn't build successfully. Given that, I've created a copy
of doc/conf.py, included in this commit, from the default doc/conf.py.in
and set the appropriate variables as necessary.

NOTE: The build didn't work when the version, release, and latex_logo
values weren't set, either directly or with interpolatable values.