Discovery: Attempt to fix crash in DiscoveryPhase::findAndCancelDeletedJob
Issue #7722

I was not able to reproduce the crash, but this is seen in the crash
reporter.

The crash occurs because the ENFORCE in that function is reached. One way
this could happen is because findAndCancelDeletedJob would be called
for the same path multiple times. Potentially this could perhaps be caused
by duplicated inodes are duplicated. But this is hard to test.