Make sure share names are in sync with server
Removing the name of a share link does not yield an error, but causes the server to automatically assign the link ID. The UI so far didn't show this, however. Rather than disallowing empty link names in the UI, we'll follow a "server makes it right" approach and trust it to either return an error or accept the request. In the latter case (which is what is experienced when emptying the name), to show up-to-date information client side, we have to re-fetch the share names from the server to show accurate information.