Use blacklist for all NormalError
The NormalError will still cause a resync after 30s.
Errors on the blacklist will be retried in the next sync, as their initial timeout (25s) is over.
If they fail again they sync will not trigger a new resync and the timeout for the affected file
will be increased to 5*25.

Fixes: #8672