Work around compiler bug in old GCC versions
The previous use of template specialisation triggered a compiler bug on
Debian 9 and CentOS 7, which use gcc 5.4 or earlier. As there is just
one specialisation, the work-around is to change the function to be a
static member function of the wrapping class of the enum.

See also: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=56480