Support dehydrated files in autotests
When creating the local state, check if files are hydrated or not.

If not hydrated, don't read the file: the OS will trigger a
download. This is bad: first the read will fail, because the test is
running on the main thread, the same place where work from callbacks
from the OS get handled. This will result in a time-out for the OS,
and it will return 0 bytes read. So the size for the file in our
local state is set to zero bytes, which makes the comparisson with
the remote state fail, which in turn makes the comparisson fail.

Worse: the callbacks from the system do come in, and are emitted as
a _QueuedConnection_. So when another call to `syncOnce` is made,
the queued downloads will be done, and the file will be re-hydrated,
thus changing the state of the files on disk.