Mac: Fix migration of launch-on-startup
Yes, we have code to use launchd plist files, but there was no code that
a) migrates the old setting into the plist file, and b) unregister the
client from the deprecated Login Items. This change fixes that
oversight.

The code here is a copy of the old code that got removed by moving to
the plist file.