Remove styling from the warning label in the "Add Folder" dialog
The styling caused unreadable warning/error messages when the OS
theme was set to dark: the text color would be white, but the
background was styled to be white too.

The message is still noticable: normally there is no text in the
label, so an error or warning message showing up will be noticed.

Fixes: https://github.com/owncloud/client/issues/11243