Fix early-use crash in folder watcher on Linux
In a few error scenarios (out-of-memory or out-of-space), the folder
watcher on Linux shows a dialog with an error message. If this occurs
during start-up before the GUI is initialised, this will lead to a
crash.

This is fixed by delaying the VFS start until after event loop is
started, guaranteeing that the GUI is up and running.

Fixes: #11460
(cherry picked from commit 8b50232a1aaee74eba8320150d72dee0597dc086)