Fix count of synced spaces
When a new account is added, the settings page for the account is
created with the account in the disconnected state. At that point, it is
not yet known if the server supports spaces, so there is no space
manager yet, and the update signal is not connected. Moving the signal
connect to `slotAccountStateChanged` for the `Connected` state fixes the
issue.

Fixes: #11889