Use standard include for QML_ELEMENT
QtQmlMeta is a weird semi-internal include that is not really
supposed to be used, and here it's only used to transitively
pull in QML_ELEMENT and friends. Qt 6.10.1 makes it no longer
public, so the build now fails. Fix it by using the include
you're supposed to be using according to Qt docs.